// Edu-Alt-Tech-website-main/types.ts
export type GlobalUserRole = 'admin';
export type UserRole = 'admin' | 'student' | 'teacher';

export interface UserObject {
  uid: string;
  name: string;
  email: string;
  role?: GlobalUserRole;
  profilePic?: string;
  createdAt: any;
  preferences?: string[]; // E.g., 'jee-mains', 'music'
  classYear?: string;
  location?: string;
}